projects/33/app/requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
asgiref==3.3.1
backoff==1.10.0
certifi==2020.11.8
CFGraph==0.2.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Django==3.1.3
Flask==1.1.2
Flask-Cors==3.0.9
geojson-rewind==0.2.0
geomet==0.2.1.post1
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonasobj==1.2.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mwoauth==0.3.7
numpy==1.19.4
oauthlib==3.1.0
pandas==1.0.5
pbr==5.5.1
Pillow==8.0.1
psycopg2==2.8.6
pygbif==0.5.0
PyJSG==0.10.0
PyJWT==1.7.1
pyparsing==2.4.7
PyShEx==0.7.14
PyShExC==0.8.2
python-dateutil==2.8.1
pytz==2020.4
rdflib==5.0.0
rdflib-jsonld==0.5.0
requests==2.24.0
requests-cache==0.5.2
requests-oauthlib==1.3.0
shexer==0.0.8
ShExJSG==0.7.0
simplejson==3.17.2
six==1.15.0
sparql-slurper==0.3.4
SPARQLWrapper==1.8.5
sqlparse==0.4.1
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
wikidataintegrator==0.8.0
pdfminer.six==20201018
nltk==3.2.5
ete3==3.1.2